UPSTREAM: [media] media: rc: if protocols can't be changed, don't be writable
If the protocols of an rc device cannot be changed, ensure the sysfs
file is not writable.

This makes it possible to detect this from userspace, so ir-keytable
can deal with case without giving an error.

Change-Id: I30a6d97ba1ef3884b44161c4aa2e6e1a5879b34a
Signed-off-by: Sean Young <sean@mess.org>
Signed-off-by: Mauro Carvalho Chehab <mchehab@osg.samsung.com>
(cherry picked from commit 68e1d3ebc480d6ed5fd0cfbea0741f07e54598c9)
Signed-off-by: z3rokwq <z3rokwq@gmail.com>