UPSTREAM: ipv6: introduce a new function fib6_update_sernum()
This function takes a route as input and tries to update the sernum in
the fib6_node this route is associated with. It will be used in later
commit when adding a cached route into the exception table under that
route.

Change-Id: I1fddfac52c688d7dd51c1ddbb1e96cf799ef4fae
Signed-off-by: Wei Wang <weiwan@google.com>
Signed-off-by: Martin KaFai Lau <kafai@fb.com>
Signed-off-by: Eric Dumazet <edumazet@google.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit 6a2bcb2c72ba01b2aac461e4d2cd39663944ebe8)
Signed-off-by: z3rokwq <z3rokwq@gmail.com>