UPSTREAM: media: lirc: allow lirc device to be opened more than once
This makes it possible for lircd to read from a lirc chardev, and not
keep it busy.

Note that this changes the default for timeout reports to on. lircd
already enables timeout reports when it opens a lirc device, leaving
them on until the next reboot.

Change-Id: Ib126fda285cbde3690f8b56ec6cac8ba23b50694
Signed-off-by: Sean Young <sean@mess.org>
Signed-off-by: Mauro Carvalho Chehab <mchehab@s-opensource.com>
(cherry picked from commit be9e7509cd00db9a2c28f1b84b1dc3eb35bcc9ad)
Signed-off-by: z3rokwq <z3rokwq@gmail.com>