UPSTREAM: bpf: revert report offload info to user space
This reverts commit bd601b6ada11 ("bpf: report offload info to user
space").  The ifindex by itself is not sufficient, we should provide
information on which network namespace this ifindex belongs to.
After considering some options we concluded that it's best to just
remove this API for now, and rework it in -next.

Change-Id: Id934c8a26171e4477ccbdf20ee38287db6b73f45
Signed-off-by: Jakub Kicinski <jakub.kicinski@netronome.com>
Signed-off-by: Daniel Borkmann <daniel@iogearbox.net>
(cherry picked from commit 97d261230bfa9e028473ca756c1625e8241eb635)
Signed-off-by: z3rokwq <z3rokwq@gmail.com>