UPSTREAM: net: Rearrange dst_entry layout to avoid useless padding.
We have padding to try and align the refcount on a separate cache
line.  But after several simplifications the padding has increased
substantially.

So now it's easy to change the layout to get rid of the padding
entirely.

We group the write-heavy __refcnt and __use with less often used
items such as the rcu_head and the error code.

Change-Id: I1e79b69ea9496846e11af0b9aeb362a6a1af54e9
Signed-off-by: David S. Miller <davem@davemloft.net>
Reviewed-by: Eric Dumazet <edumazet@google.com>
(cherry picked from commit 9481c02209617703dfcfbfba488a912f428bc2be)
Signed-off-by: z3rokwq <z3rokwq@gmail.com>