UPSTREAM: inet: Add a count to struct inet_listen_hashbucket
This patch adds a count to the 'struct inet_listen_hashbucket'.
It counts how many sk is hashed to a bucket.  It will be
used to decide if the (to-be-added) portaddr listener's hashtable
should be used during inet[6]_lookup_listener().

Change-Id: Iab869fff9a78876b78efadd2d8c08ec7a39e5d82
Signed-off-by: Martin KaFai Lau <kafai@fb.com>
Reviewed-by: Eric Dumazet <edumazet@google.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit e8964f97b82b19aa0dfbc46a878b376772247057)
Signed-off-by: z3rokwq <z3rokwq@gmail.com>