UPSTREAM: net: Allow a rule to track originating protocol
Allow a rule that is being added/deleted/modified or
dumped to contain the originating protocol's id.

The protocol is handled just like a routes originating
protocol is.  This is especially useful because there
is starting to be a plethora of different user space
programs adding rules.

Allow the vrf device to specify that the kernel is the originator
of the rule created for this device.

Change-Id: I3900a3aa07fdc4b961fe4136982044e81a95ee77
Signed-off-by: Donald Sharp <sharpd@cumulusnetworks.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit e7a2229c6e18e266492a3db9e1f40181ac2f1dae)
Signed-off-by: z3rokwq <z3rokwq@gmail.com>