net: socket: add __sys_getsockopt() helper; remove in-kernel call to syscall
Using the net-internal helper __sys_getsockopt() allows us to avoid the
internal calls to the sys_getsockopt() syscall.

This patch is part of a series which removes in-kernel calls to syscalls.
On this basis, the syscall entry path can be streamlined. For details, see
http://lkml.kernel.org/r/20180325162527.GA17492@light.dominikbrodowski.net

Cc: David S. Miller <davem@davemloft.net>
Cc: netdev@vger.kernel.org
Change-Id: Id97eddaa49de9cbc201c944ba9f3f207879c7833
Signed-off-by: Dominik Brodowski <linux@dominikbrodowski.net>
(cherry picked from commit d2827842b346507926aaad58ad6513aa68c6a574)
Signed-off-by: z3rokwq <z3rokwq@gmail.com>