BACKPORT: fs: add timespec64_truncate()
As vfs moves to using struct timespec64 to represent times,
update the argument to timespec_truncate() to use
struct timespec64. Also change the name of the function.
The rest of the implementation logic is the same.

Move this to fs/inode.c instead of kernel/time/time.c as all the
users of this api are filesystems.

Change-Id: I19b58d7d1d5ec263b04237a33b8f981e7d407171
Signed-off-by: Deepa Dinamani <deepa.kernel@gmail.com>
Cc: <viro@zeniv.linux.org.uk>
(cherry picked from commit 057149d47c7faaa90a4d1122bd10ce6de9bc78d0)
Signed-off-by: z3rokwq <z3rokwq@gmail.com>