UPSTREAM: xdp: add MEM_TYPE_ZERO_COPY
Here, a new type of allocator support is added to the XDP return
API. A zero-copy allocated xdp_buff cannot be converted to an
xdp_frame. Instead is the buff has to be copied. This is not supported
at all in this commit.

Also, an opaque "handle" is added to xdp_buff. This can be used as a
context for the zero-copy allocator implementation.

Change-Id: I96e40ee36d27ff69d7dcb4600eb608d1d053e451
Signed-off-by: Björn Töpel <bjorn.topel@intel.com>
Signed-off-by: Daniel Borkmann <daniel@iogearbox.net>
(cherry picked from commit f952041e5d9bb391f98978dc8c019051025619b4)
Signed-off-by: z3rokwq <z3rokwq@gmail.com>